fix: cambio a float

+++ Clases/src/com/krakedev/Plato.java
@@ -4,7 +4,7 @@ public class Plato {
 
 	public String nombre;
 	public String tipo;
-	public double precio;
+	public float precio;
 	public boolean disponible;
 	
 }
