+++ Clases/src/com/krakedev/Restaurante.java
@@ -4,6 +4,6 @@ public class Restaurante {
 
 	public String nombre;
 	public String direccion;
-	public double calificacion;
+	public float calificacion;
 	
 }
